chore: remove anonymous usage stats
Taken out at the user's request. The app no longer counts anything, sends
anything, or carries a setting about it.

Client:
- `usage.rs` deleted, with every `usage::track` call site across `main`,
  `overlay` and `voice::session`, the `track_event` and
  `set_analytics_enabled` commands, and the startup hook
- `analytics_enabled` and `install_id` off `AppConfig` — the install id existed
  only to key the reports
- `lib/analytics.ts` deleted, the Dashboard's page-name derivation and its two
  counting effects with it, and the Welcome screen's note
- the Settings toggle, and `settings.analytics`, `settings.analyticsDesc` and
  `welcome.analyticsNote` from all six locales

Control plane:
- `usage.ts` and `usagepage.ts` deleted; `POST /v1/usage`,
  `GET /v1/usage/stats`, the `/admin/usage` dashboard and the prune from the
  cron all go with them
- `adminAllowed` and `windowDays` were living in `usage.ts` but are the gate on
  every admin page, so they move to `auth.ts` beside `tokenMatches` rather than
  leaving with the feature. Their tests move to `admingate.test.ts`
- `/admin` and `/admin/usage` now redirect to the first remaining tab, so an
  old bookmark still lands somewhere real

Not done, deliberately: `usage_daily` and `usage_events` are left in place.
Dropping a table is destructive and those hold what was already collected, so
whether they go is a separate decision. Nothing writes to them now.

1,477 Rust tests and 275 control-plane tests pass; typecheck and the production
build are clean.
