+++ apps/manager/src/i18n/locales/de.ts
@@ -130,8 +130,6 @@ export const de: Translation = {
   "welcome.intro.body":
     "Dein Mod-Manager für MX Bikes. Halte Strecken, Motorräder und Lackierungen an einem Ort organisiert — keine ZIP-Dateien mehr über den ganzen Desktop verstreut. Wir zeigen dir das Wichtigste in ein paar Sekunden.",
   "welcome.getStarted": "Los geht's",
-  "welcome.analyticsNote":
-    "{{app}} zählt anonym, welche Seiten genutzt werden. Keine Namen, keine Dateien — in den Einstellungen abschaltbar.",
   "presets.missingMods":
     "Fehlende Mods: {{mods}}. Installiere sie, damit diese Teile angezeigt werden.",
   "presets.help":
@@ -608,9 +606,6 @@ export const de: Translation = {
   "settings.colorwayRetro": "Retro",
   "settings.language": "Sprache",
   "settings.languageSystem": "System",
-  "settings.analytics": "Anonyme Nutzungsstatistik teilen",
-  "settings.analyticsDesc":
-    "Sendet eine zufällige ID, die App-Version und Zähler dazu, welche Seiten und Funktionen du nutzt — damit klar wird, was sich zu bauen lohnt. Niemals dein Name, deine Dateien oder deine Adresse.",
   "settings.runInBackground": "Im Hintergrund weiterlaufen",
   "settings.runInBackgroundDesc":
     "Beim Schließen des Fensters läuft {{app}} im Infobereich weiter, damit FrostMod verbunden bleibt. Beenden über das Symbol im Infobereich.",
