+++ apps/manager/src/i18n/locales/en.ts
@@ -123,8 +123,6 @@ export const en = {
   "welcome.intro.body":
     "Your mod manager for MX Bikes. Keep your tracks, bikes and paints organized in one place — no more zip files scattered across your desktop. We'll show you around in a few seconds.",
   "welcome.getStarted": "Get started",
-  "welcome.analyticsNote":
-    "{{app}} counts which pages get used, anonymously. No names, no files — you can turn it off in Settings.",
   "presets.missingMods":
     "Missing mods: {{mods}}. Install them for those parts to show.",
   "presets.help":
@@ -594,9 +592,6 @@ export const en = {
   "settings.colorwayRetro": "Retro",
   "settings.language": "Language",
   "settings.languageSystem": "System",
-  "settings.analytics": "Share anonymous usage stats",
-  "settings.analyticsDesc":
-    "Sends a random ID, the app version and counts of which pages and features you use, so I can tell what's worth building. Never your name, your files or your address.",
   "settings.runInBackground": "Keep running in the background",
   "settings.runInBackgroundDesc":
     "Closing the window hides {{app}} to the tray so FrostMod stays connected. Quit from the tray icon.",
